server/requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.2.1
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.4
cryptography==38.0.3
decorator==5.1.0
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.13
django-admin==2.0.2
django-allauth==0.50.0
django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-excel-response2==3.0.5
django-extensions==3.1.3
django-seed==0.3.1
django-six==1.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
excel-base==1.0.4
executing==1.2.0
Faker==13.3.4
idna==3.4
inflection==0.5.1
ipython==7.27.0
isoweek==1.3.3
jedi==0.18.0
jsonschema==4.17.0
matplotlib-inline==0.1.3
oauthlib==3.2.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.20
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.10.0
PyJWT==2.6.0
pyrsistent==0.19.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
screen==1.0.1
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
stack-data==0.6.2
TimeConvert==3.0.13
toposort==1.7
traitlets==5.1.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
xlwt==1.3.0
